worker/pyproject.toml
[project]
name = "hydra-worker"
version = "0.1.0"
description = "GPU worker for Hydra distributed LLM inference"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML
    "torch>=2.2.0",
    "safetensors>=0.4.0",
    "tokenizers>=0.15.0",
    "transformers>=4.36.0",

    # Communication
    "pyzmq>=25.0.0",
    "msgpack>=1.0.0",

    # Utilities
    "numpy>=1.24.0",
    "psutil>=5.9.0",
    "structlog>=23.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
training = [
    "bitsandbytes>=0.42.0",
    "peft>=0.7.0",
    "wandb>=0.16.0",
    "accelerate>=0.25.0",
]
flash = [
    "flash-attn>=2.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
hydra-worker = "hydra_worker.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
